Se agregaron Productos vendidos y lista ventas

+++ WormBook/Models/Productosenviado.cs
@@ -7,7 +7,7 @@ public partial class Productosenviado
 {
     public int CodigoInterno { get; set; }
 
-    public int Guiaenvio { get; set; }
+    public long Guiaenvio { get; set; }
 
     public int Cantidadenviada { get; set; }
     public virtual Producto CodigoInternoNavigation { get; set; } = null!;
